requirements.txt
streamlit>=1.28.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
httpx>=0.25.0
